refactor: move component extracts inside openedx folder
