+++ backend/openedx_ai_extensions/processors/openedx/utils/component_extractors.py
